% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.shieh.d.R
\name{tr.shieh.d}
\alias{tr.shieh.d}
\title{tr.shieh.d function to compute transformed Shieh's d and its confidence intervals}
\usage{
tr.shieh.d(x, y, conf.level, alternative, na.rm)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values (first sample)}

\item{y}{a (non-empty) numeric vector of data values (second sample)}

\item{conf.level}{confidence level of the interval}

\item{alternative}{a character string describing the alternative hypothesis (two.sided, less or greater)}

\item{na.rm}{set whether Missing Values should be excluded (na.rm = TRUE) or not (na.rm = FALSE) - defaults to TRUE}
}
\value{
Returns Call, Shieh's d, confidence interval, sample size and sample estimates
}
\description{
tr.shieh.d function to compute transformed Shieh's d and its confidence intervals
}
\keyword{Shieh's}
\keyword{Two}
\keyword{comparison,}
\keyword{confidence}
\keyword{d,}
\keyword{effect}
\keyword{interval}
\keyword{means}
\keyword{size,}
